Remove CAPO/OpenStack support from all commands (#1728)
* Remove CAPO/OpenStack support from all commands

* Update CHANGELOG.md
