+++ cmd/gitops/add/base/command.go
@@ -27,9 +27,7 @@ https://github.com/giantswarm/gitops-template/blob/main/docs/repo_structure.md.`
 
 	examples = `  # Add a CAPA base to your GitOps directory structure
   kubectl gs gitops add base --provider capa
-
-  # Add a CAPO base to your GitOps directory structure
-  kubectl gs gitops add base --provider openstack`
+`
 )
 
 type Config struct {
