+++ cmd/login/wc.go
@@ -109,7 +109,7 @@ func (r *runner) getCluster(ctx context.Context, services serviceSet, provider s
 }
 
 func (r *runner) getCertOperatorVersion(c *cluster.Cluster, provider string, services serviceSet, ctx context.Context) (string, error) {
-	// Pure CAPI providers (e.g. openstack) do not have release CRs and labels nor do they run cert-operator. so we return an empty string.
+	// Pure CAPI providers do not have release CRs and labels nor do they run cert-operator. so we return an empty string.
 	// In this case wecreate the client certificate using the MC PKI
 	if key.IsPureCAPIProvider(provider) {
 		return "", nil
