--- cmd/template/cluster/provider/templates/openstack/functions.go
@@ -1,52 +1,0 @@
-package openstack
-
-import (
-	"github.com/giantswarm/microerror"
-	"sigs.k8s.io/yaml"
-)
-
-func GenerateClusterValues(flagInputs ClusterConfig) (string, error) {
-	var flagConfigData map[string]interface{}
-
-	{
-		flagConfigYAML, err := yaml.Marshal(flagInputs)
-		if err != nil {
-			return "", microerror.Mask(err)
-		}
-
-		err = yaml.Unmarshal(flagConfigYAML, &flagConfigData)
-		if err != nil {
-			return "", microerror.Mask(err)
-		}
-	}
-
-	finalConfigString, err := yaml.Marshal(flagInputs)
-	if err != nil {
-		return "", microerror.Mask(err)
-	}
-
-	return string(finalConfigString), nil
-}
-
-func GenerateDefaultAppsValues(flagConfig DefaultAppsConfig) (string, error) {
-	var flagConfigData map[string]interface{}
-
-	{
-		flagConfigYAML, err := yaml.Marshal(flagConfig)
-		if err != nil {
-			return "", microerror.Mask(err)
-		}
-
-		err = yaml.Unmarshal(flagConfigYAML, &flagConfigData)
-		if err != nil {
-			return "", microerror.Mask(err)
-		}
-	}
-
-	finalConfigString, err := yaml.Marshal(flagConfigData)
-	if err != nil {
-		return "", microerror.Mask(err)
-	}
-
-	return string(finalConfigString), nil
-}
