+++ cmd/template/cluster/runner.go
@@ -98,11 +98,6 @@ func (r *runner) run(ctx context.Context, client k8sclient.Interface) error {
 		if err != nil {
 			return microerror.Mask(err)
 		}
-	case key.ProviderOpenStack:
-		err = provider.WriteOpenStackTemplate(ctx, client, output, config)
-		if err != nil {
-			return microerror.Mask(err)
-		}
 	case key.ProviderVSphere:
 		err = provider.WriteVSphereTemplate(ctx, client, output, config)
 		if err != nil {
@@ -139,7 +134,6 @@ func (r *runner) getClusterConfig() (common.ClusterConfig, error) {
 		AWS:           r.flag.AWS,
 		Azure:         r.flag.Azure,
 		OIDC:          r.flag.OIDC,
-		OpenStack:     r.flag.OpenStack,
 		VSphere:       r.flag.VSphere,
 		CloudDirector: r.flag.CloudDirector,
 	}
