+++ internal/feature/providers.go
@@ -3,9 +3,8 @@ package feature
 const (
 	ProviderAWS           = "aws"
 	ProviderAzure         = "azure"
-	ProviderOpenStack     = "openstack"
 	ProviderCAPA          = "capa"
 	ProviderCAPZ          = "capz"
 	ProviderCloudDirector = "cloud-director"
-	ProviderVSphere       = "vspshere"
+	ProviderVSphere       = "vspshere" // TODO: fix spelling
 )
