+++ internal/gitops/filesystem/creator/testdata/expected/create/2-kustomization.golden
@@ -3,4 +3,4 @@ commonLabels:
   giantswarm.io/managed-by: flux
 kind: Kustomization
 resources:
-- ../../../../../../../bases/cluster/capo
+- ../../../../../../../bases/cluster/capa
