+++ internal/gitops/filesystem/creator/testdata/expected/dry-run/case-2-flawless.golden
@@ -31,4 +31,4 @@ commonLabels:
   giantswarm.io/managed-by: flux
 kind: Kustomization
 resources:
-- ../../../../../../../bases/cluster/capo
+- ../../../../../../../bases/cluster/capa
