+++ internal/gitops/filesystem/creator/testdata/input/1-kustomization.yaml
@@ -3,4 +3,4 @@ commonLabels:
   giantswarm.io/managed-by: flux
 kind: Kustomization
 resources:
-- ../../../../../../../bases/cluster/capo
+- ../../../../../../../bases/cluster/capa
