+++ internal/gitops/structure/workload-cluster/testdata/expected/0-kustomization.golden
@@ -3,4 +3,4 @@ commonLabels:
   giantswarm.io/managed-by: flux
 kind: Kustomization
 resources:
-  - ../../../../../../../bases/cluster/capo
+  - ../../../../../../../bases/cluster/capa
